// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  user: 'user',
  partner: 'partner',
  admin: 'admin'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const PartnerStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected'
} as const

export type PartnerStatus = (typeof PartnerStatus)[keyof typeof PartnerStatus]


export const VideoKycStatus = {
  not_required: 'not_required',
  pending: 'pending',
  inprogress: 'inprogress',
  approved: 'approved',
  rejected: 'rejected'
} as const

export type VideoKycStatus = (typeof VideoKycStatus)[keyof typeof VideoKycStatus]


export const VehicleType = {
  bike: 'bike',
  car: 'car',
  loading: 'loading',
  truck: 'truck',
  auto: 'auto'
} as const

export type VehicleType = (typeof VehicleType)[keyof typeof VehicleType]


export const ApprovalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected'
} as const

export type ApprovalStatus = (typeof ApprovalStatus)[keyof typeof ApprovalStatus]


export const BankStatus = {
  not_added: 'not_added',
  added: 'added',
  verified: 'verified'
} as const

export type BankStatus = (typeof BankStatus)[keyof typeof BankStatus]


export const BookingStatus = {
  idle: 'idle',
  requested: 'requested',
  awaiting_payment: 'awaiting_payment',
  confirmed: 'confirmed',
  started: 'started',
  completed: 'completed',
  cancelled: 'cancelled',
  rejected: 'rejected',
  expired: 'expired'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const PaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  cash: 'cash',
  failed: 'failed'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const ChatSender = {
  user: 'user',
  driver: 'driver'
} as const

export type ChatSender = (typeof ChatSender)[keyof typeof ChatSender]


export const PayoutMethod = {
  bank: 'bank',
  upi: 'upi'
} as const

export type PayoutMethod = (typeof PayoutMethod)[keyof typeof PayoutMethod]


export const PayoutStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  rejected: 'rejected'
} as const

export type PayoutStatus = (typeof PayoutStatus)[keyof typeof PayoutStatus]